+++ build-config/scripts/add-banner.js
@@ -4,21 +4,20 @@
  * @file add-banner.js
  * @since 2.0.0
  */
+
 var fs = require('fs');
 var path = require('path');
-var moment = require('moment');
 var banner = require('add-banner');
 var argvAutoGlob = require('argv-auto-glob');
 
 var CSS = '.css';
-var JS = '.js';
 var infile, fpath, result;
 var pjson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
 
 // check command-line arguments
 process.argv = argvAutoGlob(process.argv);
 if (process.argv.length < 3) {
-    console.error('Error: no input file(s) specified');
+    console.error("Error: no input file(s) specified");
     process.exit();
 }
 
@@ -28,19 +27,11 @@ process.argv.slice(2).forEach(function(fpath) {
     if (fpath.endsWith(CSS)) {
         // CSS
         result = banner(infile, {
-            banner: path.resolve('scripts', 'banner.css'),
+            banner: path.resolve(__dirname, 'banner.css'),
             pkg: pjson
         });
         fs.writeFile(fpath, result);
 
-    } else if (fpath.endsWith(JS)) {
-        // javascript
-        result = banner(infile, {
-            banner: path.resolve('scripts', 'banner.ejs'),
-            pkg: pjson,
-            moment: moment
-        });
-        fs.writeFile(fpath, result);
     }
     console.info('Added banner to ' + fpath + ' with version ' + pjson.version);
 });