+++ src/js/videojs.record.js
@@ -864,7 +864,7 @@ class Record extends Plugin {
                 this._processing = false;
 
                 // hide loader
-                this.player.loadingSpinner.hide( );
+                this.player.loadingSpinner.hide();
 
                 // show animation total duration
                 this.setDuration(this.streamDuration);
@@ -1227,8 +1227,13 @@ class Record extends Plugin {
                 try {
                     var track = this.stream.getVideoTracks()[0];
                     var imageCapture = new ImageCapture(track);
+                    let photoSettings = {
+                        imageWidth: recordCanvas.width,
+                        imageHeight: recordCanvas.height
+                    };
 
-                    imageCapture.takePhoto().then((blob) => {
+                    // take picture
+                    imageCapture.takePhoto(photoSettings).then((blob) => {
                         return createImageBitmap(blob);
 
                     }).then((imageBitmap) => {
@@ -1450,12 +1455,14 @@ class Record extends Plugin {
     }
 }
 
-// version nr gets replaced during build
-Record.VERSION = 'dev';
+// version nr is injected during build
+Record.VERSION = __VERSION__;
 
 // register plugin
 videojs.Record = Record;
-videojs.registerPlugin('record', Record);
+if (videojs.getPlugin('record') === undefined) {
+    videojs.registerPlugin('record', Record);
+}
 
 // export plugin
 module.exports = {